src/main/java/codesquad/user/auth/AdminUserHandlerMethodArgumentResolver.java
package codesquad.user.auth;

import codesquad.user.domain.User;
import codesquad.user.exception.UnAuthorizedException;
import codesquad.util.SessionUtil;
import org.springframework.core.MethodParameter;
import org.springframework.web.bind.support.WebDataBinderFactory;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.method.support.HandlerMethodArgumentResolver;
import org.springframework.web.method.support.ModelAndViewContainer;

public class AdminUserHandlerMethodArgumentResolver implements HandlerMethodArgumentResolver {

    @Override
    public boolean supportsParameter(MethodParameter parameter) {
        return parameter.hasParameterAnnotation(AdminUser.class);
    }

    @Override
    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
        User user = SessionUtil.getUserFromSession(webRequest);

        if (user == null || !user.isAdmin()) {
            throw new UnAuthorizedException();
        }

        return user;
    }
}
